# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API User",
        "description": "Esta API é uma Basica consulta de Read,Create,Update e Delete de um banco de Usuários",
        "contact": {
            "email": "lucassantosnacs.21@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API de test"
        }
    ],
    "paths": {
        "/logar": {
            "post": {
                "summary": "entrar com email e senha para gerar seu Token",
                "description": "Essa rota é responsavel por criar um token para permitir as outras requisições",
                "tags": [
                    "authorization"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Users"
                            },
                            "examples": {
                                "User": {
                                    "value": {
                                        "email": "exemple123@g.com",
                                        "password": "senhaExemplo123"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "email ou senha invalidos"
                    },
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/newuser": {
            "post": {
                "summary": "Cadastro de Usuários",
                "description": "Essa rota é responsavel por criar novos usuarios",
                "tags": [
                    "newUsers"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Users"
                            },
                            "examples": {
                                "User": {
                                    "value": {
                                        "name": "Exemple",
                                        "email": "exemple123@g.com",
                                        "state": "Estado(sigla)",
                                        "password": "senhaExemplo123",
                                        "photo": "https://i.pinimg.com/564x/8a/e9/e9/8ae9e92fa4e69967aa61bf2bda967b7b.jpg"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Não Autorizado"
                    },
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Retorno de Usuários",
                "description": "Essa rota é responsavel por Trazer Todos os dados do banco",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "summary": "Busca ID",
                "description": "Pega um unico Usuario pelo seu ID",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Usuario para busca",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Atualizar User",
                "description": "Esta rota é responsavel por Atualizar Usuario",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Usuario para busca",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Users"
                            },
                            "examples": {
                                "User": {
                                    "value": {
                                        "name": "Antony Stark",
                                        "email": "tonyStark2008@gmail.com",
                                        "state": "Manhattan",
                                        "password": "I_am_Iro_Man2008",
                                        "photo": "https://i.pinimg.com/564x/8a/e9/e9/8ae9e92fa4e69967aa61bf2bda967b7b.jpg"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deleta User",
                "description": "Esta rota é responsavel por deletar um Usuario",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Usuario para busca",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Users": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "photo": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}